chore: move client and options to /lib
