+++ packages/cli/src/resources/broadcasts.ts
@@ -2,9 +2,9 @@
 
 import { Command } from 'commander';
 
-import { getClient } from '../client';
+import { getClient } from '../lib/client';
+import { parseOptions } from '../lib/options';
 import { printJson } from '../lib/printer';
-import { parseOptions } from '../options';
 import { broadcastsNotifications } from './broadcasts/notifications';
 
 export const broadcasts = new Command('broadcasts').description('Manage broadcasts');
